requirements.txt
ipykernel==6.21.1
ipython==8.9.0
pandas==1.5.3
scipy==1.10.0
tqdm==4.64.1
gitpython==3.1.30
segmentation-models-pytorch==0.3.2
scikit-image==0.19.3
requests==2.28.2
scikit-learn==1.2.0
imagecodecs==2023.1.23
pytorch-lightning==1.9.1
wandb==0.13.10
